docs(wip): refactor demo
