+++ demo/colors.ts
@@ -1,0 +1,8 @@
+export enum Colors {
+  primary = "dodgerblue",
+  secondary = "#3f4c6b",
+  tertiary = "#4f85bb",
+  black = "#2B2D27",
+  background = "#fcfff4",
+  transparent = "rgba(252, 255, 244, 0.8)",
+}
