+++ demo/docs/support.md
@@ -1,2 +1,4 @@
+# Support
+
 This project is free and open-source, so if you think this project can help you or anyone else, you should star it in github  
 Also feel free to open an issue if you have any idea, question, or you've found a bug. Any feedback is good support
